# openapi.yaml
openapi: 3.0.3
info:
  title: Product REST Service
  version: 1.0.0
  description: JAX-RS lab API for listing, retrieving, and creating products.
servers:
  - url: http://localhost:8080/product-rest-service/api
paths:
  /products:
    get:
      summary: List products (paged)
      parameters:
        - {name: page, in: query, schema: {type: integer, minimum: 0, default: 0}}
        - {name: size, in: query, schema: {type: integer, minimum: 1, maximum: 100, default: 10}}
      responses:
        '200': {description: Paged collection, content: {application/json: {schema: {$ref: '#/components/schemas/PagedProducts'}}}}
        '400': {$ref: '#/components/responses/BadRequest'}
    post:
      summary: Create a product
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/Product'}}}
      responses:
        '201': {description: Created, headers: {Location: {description: URL of created product, schema: {type: string, format: uri}}}, content: {application/json: {schema: {$ref: '#/components/schemas/Product'}}}}
        '400': {$ref: '#/components/responses/BadRequest'}
        '409': {description: Duplicate ID, content: {application/problem+json: {schema: {$ref: '#/components/schemas/Problem'}}}}
  /products/{id}:
    get:
      summary: Get product by ID
      parameters:
        - {name: id, in: path, required: true, schema: {type: integer, format: int64, minimum: 1}}
      responses:
        '200': {description: Product found, content: {application/json: {schema: {$ref: '#/components/schemas/Product'}}}}
        '400': {$ref: '#/components/responses/BadRequest'}
        '404': {description: Product not found, content: {application/problem+json: {schema: {$ref: '#/components/schemas/Problem'}}}}
components:
  responses:
    BadRequest: {description: Invalid input, content: {application/problem+json: {schema: {$ref: '#/components/schemas/Problem'}}}}
  schemas:
    Product:
      type: object
      required: [id, name, price, stock]
      properties:
        id: {type: integer, format: int64, minimum: 1, example: 3}
        name: {type: string, minLength: 1, maxLength: 100, example: Keyboard}
        description: {type: string, maxLength: 500, nullable: true, example: USB keyboard}
        price: {type: number, format: double, exclusiveMinimum: true, minimum: 0, example: 1299.00}
        stock: {type: integer, minimum: 0, example: 25}
    PagedProducts:
      type: object
      required: [items, page, size, totalItems, totalPages]
      properties:
        items: {type: array, items: {$ref: '#/components/schemas/Product'}}
        page: {type: integer}
        size: {type: integer}
        totalItems: {type: integer}
        totalPages: {type: integer}
    Problem:
      type: object
      properties:
        type: {type: string}
        title: {type: string}
        status: {type: integer}
        detail: {type: string}
        instance: {type: string}
        timestamp: {type: string, format: date-time}
        errors: {type: array, items: {type: string}}