# src/main/java/com/harsh/productapi/model/Product.java
package com.harsh.productapi.model;
import java.math.BigDecimal;
public class Product {
 private Long id; private String name; private String description; private BigDecimal price; private Integer stock;
 public Product(){}
 public Product(Long id,String name,String description,BigDecimal price,Integer stock){this.id=id;this.name=name;this.description=description;this.price=price;this.stock=stock;}
 public Long getId(){return id;} public void setId(Long id){this.id=id;}
 public String getName(){return name;} public void setName(String name){this.name=name;}
 public String getDescription(){return description;} public void setDescription(String description){this.description=description;}
 public BigDecimal getPrice(){return price;} public void setPrice(BigDecimal price){this.price=price;}
 public Integer getStock(){return stock;} public void setStock(Integer stock){this.stock=stock;}
}